UserManagement/UserManagement_Server/src/requirements.txt
fastapi
passlib
pydantic
pyJWT
psycopg2
uvicorn
jsonrpcserver
abc